Working on local ~ Bus Commit
Reasonably confident this will do the trick. Reasonably confident that Karate documentation is a waking nightmare.

Anyway, the kicker seems to be that by naming it 'KarateTest' it was inadvertently getting caught up with Junit tests because it matched the common naming structure for those tests (something to do with surefire reports). We'll see what happens when I'm back online.