Better data practices
Crafted a selection of assets in an isolated workflow (so as not to break existing functionality) and ensure I'm hitting all the right points.

Rather than actually delivering the model that is driving behind-the-scenes relationships, I'm instead going to serve up a DAO that I can structure deliberately. This is a simple recreation of what's being done on the model currently (with logic sent to the DAO instead). 

Leveraged Command-Line-Runner to see it in action and then ultimately created a controller to review in the browser.