PokemonDao on paged responses
Figured out a way to transform from the repository object to a DAO via mapping on the pageable object.

I set up the pokemon/{id} endpoint to return back the raw object for debugging purposes. Evolution still needs to be sorted out as it's kind of in a weird spot right now, but types are looking good.