Shiny support
Using JPA to allow for the addition of a shiny param to the query. Used an optional boolean so there can be 3 different choices (essentially true, false, null) which works fine. Karate tests included for this functionality, with a slight upgrade to how they run (see readme) to account for unicode characters in the nidorans.