Gen support with integration tests
Some existing methods on the repository were shifted around. The if statement that drives the selection of the repository method is sufficient for now, but will quickly spiral out of control if more parameters are added.

Given that this is all v1 calls for, I'm going to stick with it for now. Ideally the integration tests will be very helpful for any refactoring that comes in the future.

This should complete the pokemon side of things for v1, barring any additional sorting logic.