Box 0.2.0 bugfix
There was a database container running on the initial box that was causing changes to not always appear. With this box update, it will be an actual fresh image so database changes can propagate.

In addition I started migrating startup scripts into the dev/ directory to avoid muddying up the root of the repo.