Moves Join Table
Pulled evolutions, tired of looking at it.

pkmn_mvs joins pokemon to their moves and is generated the same as all the other .sql scripts. I tested that the database does indeed startup with these script values, so as far as I can see, this should be all set for moving onto the API side of things.