Agent label
This coincides with Jenkins EC2 plugin configuration which will spin up a t2.micro EC2 instance with the label 'stage'. No preloading on this instance beyond what is required for the slave.jar to run, so next up is iteratively making sure it's configured to include all the necessary dependencies.

(Also using this commit to double-check the github webhook)